// EigenDA/src/types/schema.ts
export class Batch {
  static entity = "Batch";
  static schema = {
    id: { type: "String", index: true },
    batchId: "string",
    batchHeaderHash: "string",
    gasFees: "string",
    blockNumber: "string",
    blockTimestamp: "string",
    transactionHash: "string",
    batchHeader: {
      id: "String",
      blobHeadersRoot: "string",
      quorumNumbers: "number[]",
      signedStakeForQuorums: "number[]",
      referenceBlockNumber: "number",
    },
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class GasFees {
  static entity = "GasFees";
  static schema = {
    id: { type: "String", index: true },
    gasUsed: "string",
    gasPrice: "string",
    transactionFee: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}